feat: move CustomerPortalForm to a new location and update subscription cancellation logic
