+++ app/account/page.tsx
@@ -1,4 +1,3 @@
-import CustomerPortalForm from '@/components/ui/AccountForms/CustomerPortalForm';
 import EmailForm from '@/components/ui/AccountForms/EmailForm';
 import NameForm from '@/components/ui/AccountForms/NameForm';
 import { redirect } from 'next/navigation';
@@ -8,6 +7,7 @@ import {
   getSubscription,
   getUser
 } from '@/utils/supabase/queries';
+import CustomerPortalForm from '@/components/ui/AccountForms/CustomerPortalForm';
 
 export default async function Account() {
   const supabase = createClient();
